+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 * Save and close batch allocations. Refs UIF-534.
 * Add batch allocation log details screen. Refs UIF-531.
 * Search and filter batch allocation logs. Refs UIF-541.
+* Disable 'Save & close' button on Batch edit screen if changes were made after clicking 'Recalculate' button. Refs UIF-591.
+* Move View batch allocation logs action to list level. Refs UIF-588.
 * *BREAKING* Migrate stripes dependencies to their Sunflower versions. Refs UIF-574.
 * *BREAKING* Migrate `react-intl` to v7. Refs UIF-575.